Cleaned up a little
